Reworked the datafile, and logic around rendering to remove unnecesarry dependencies on state values
In this update, I changed my SliderData to be an array of objects instead of and object of arrrays so that I could easily map images to display. In this update I added functionality to the menu so that you could pick a folder and have it display in the app. Next, I am going to add some more content to the about, work on styles, image-sizing, and add a title on the page.